Refactor product detail folder so reviews make more sense closes #95

+++ server/db/models/review.js
@@ -5,12 +5,12 @@ module.exports = function (db) {
 
     var Review = db.define('review', {
        stars: {
-       	type: Sequelize.INTEGER,
+       	type: Sequelize.DECIMAL(10,1),
        	allowNull: false,
         validate: {
           isInt: true,
-          min: 1,
-          max: 5
+          min: 1.0,
+          max: 5.0
         }
        },
